+++ 106/InstaCat/InstaCat/.swiftlint.yml
