+++ 106/InstaCat/InstaCat/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  InstaCat
+//
+//  Created by Камил Хайрутдинов on 28.10.2023.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .red
+        // Do any additional setup after loading the view.
+    }
+
+}
+
